+++ tests/testthat/test-translate.r
@@ -1,11 +1,5 @@
 context("translate")
 
-test_that("logical values are converted to 0/1/NULL", {
-  expect_equal(translate_sql(FALSE), sql("0"))
-  expect_equal(translate_sql(TRUE), sql("1"))
-  expect_equal(translate_sql(NA), sql("NULL"))
-})
-
 test_that("dplyr.strict_sql = TRUE prevents auto conversion", {
   old <- options(dplyr.strict_sql = TRUE)
   on.exit(options(old))